Numpy v2+ update and MoonMag integration: Update PP to use be compatible with numpy v2. Part of this integration moves MoonMag into the PP repository, as MoonMag used data types taht are no longer part of numpy v2 (i..e complex_). Thus we have integrated MoonMag completely into PlanetProfile which allows us to update MoonMag as necessary.

+++ PlanetProfile/Thermodynamics/HydroEOS.py
@@ -1080,13 +1080,13 @@ class SFphase:
                 PT = sfPTgrid(P_MPa, T_K)
             else:
                 PT = self.PTpairs(P_MPa, T_K)
-            return WhichPhase(PT).astype(np.int_)
+            return WhichPhase(PT).astype(np.int8)
         else:
             if grid:
                 PTm = sfPTmGrid(P_MPa, T_K, self.m_molal)
             else:
                 PTm = self.PTmTrips(P_MPa, T_K)
-            return WhichPhase(PTm, solute=self.comp).astype(np.int_)
+            return WhichPhase(PTm, solute=self.comp).astype(np.int8)
         
 
 class SFSeismic:
